Extract invitation handling
