+++ src/pages/purchase-orders/edit/Edit.tsx
@@ -18,7 +18,7 @@ import { useResolveCurrency } from 'common/hooks/useResolveCurrency';
 import { useTitle } from 'common/hooks/useTitle';
 import { useVendorResolver } from 'common/hooks/vendors/useVendorResolver';
 import { InvoiceItem } from 'common/interfaces/invoice-item';
-import { Invitation, PurchaseOrder } from 'common/interfaces/purchase-order';
+import { PurchaseOrder } from 'common/interfaces/purchase-order';
 import { ValidationBag } from 'common/interfaces/validation-bag';
 import { usePurchaseOrderQuery } from 'common/queries/purchase-orders';
 import { blankLineItem } from 'common/stores/slices/invoices/constants/blank-line-item';
@@ -44,6 +44,7 @@ import { generatePath, useParams } from 'react-router-dom';
 import { v4 } from 'uuid';
 import { Details } from './components/Details';
 import { VendorSelector } from './components/VendorSelector';
+import { useHandleInvitationChange } from './hooks/useHandleInvitationChange';
 
 export function Edit() {
   const { documentTitle } = useTitle('edit_purchase_order');
@@ -87,29 +88,7 @@ export function Edit() {
     setPurchaseOrder((current) => current && { ...current, [property]: value });
   };
 
-  const handleInvitationChange = (id: string, checked: boolean) => {
-    let invitations = [...purchaseOrder!.invitations];
-
-    const potential =
-      invitations?.find((invitation) => invitation.vendor_contact_id === id) ||
-      -1;
-
-    if (potential !== -1 && checked === false) {
-      invitations = invitations.filter((i) => i.vendor_contact_id !== id);
-    }
-
-    if (potential === -1) {
-      const invitation: Partial<Invitation> = {
-        vendor_contact_id: '',
-      };
-
-      invitation.vendor_contact_id = id;
-
-      invitations.push(invitation as Invitation);
-    }
-
-    handleChange('invitations', invitations);
-  };
+  const handleInvitationChange = useHandleInvitationChange(handleChange);
 
   const resolveCurrency = async (vendorId: string) => {
     const vendor = await vendorResolver.find(vendorId);
@@ -191,7 +170,9 @@ export function Edit() {
           resource={purchaseOrder}
           onChange={(id) => handleChange('vendor_id', id)}
           onClearButtonClick={() => handleChange('vendor_id', '')}
-          onContactCheckboxChange={handleInvitationChange}
+          onContactCheckboxChange={(id, checked) =>
+            purchaseOrder && handleInvitationChange(purchaseOrder, id, checked)
+          }
         />
 
         <Details
